demo-python/code/embeddings/cohere-embeddings/cohere-embeddings-requirements.txt
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
azure-identity==1.16.0
aiohttp
ipykernel
python-dotenv
cohere
azure-ai-ml
azure-mgmt-storage
azure-mgmt-search